chore: resolve SectionHeader merge conflicts
