+++ src/app/user/Communities/CommunitiesSection.tsx
@@ -2,8 +2,7 @@ import { CommunityCard } from '@/app/user/Communities/CommunityCard'
 import { JoinACommunity } from '@/app/user/Communities/JoinACommunity'
 import { useCommunity } from '@/shared/hooks/useCommunity'
 import { communitiesMapByContract } from '@/app/communities/communityUtils'
-import { SectionHeader } from '@/components/section-header'
-import { Link } from '@/components/Link'
+import { SectionHeader } from '@/components/SectionHeader'
 import { useAccount } from 'wagmi'
 import { useEffect, useRef, useState } from 'react'
 import { useNFTBoosterContext } from '@/app/providers/NFT/BoosterContext'
